added abstract class AbstractBaseTest
